app/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
httpx
pydantic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
